Rework CI detection logic
Our existing logic was ported from an existing NPM package. However,
our port was incorrect, which is unsurprising given the style of the
source code.

I have rewritten the logic to be more idomatic golang, while keeping
the overall detection logic intact.

Our detection is still based on the presence of environment
variables.

Fixes: #1646