Fix pulumi/pulumi#1995 (#1997)
The diff display code was not expecting that it would be possible for
resource properties to transition from being an archive to being an
asset, or the other way around. This commit prints out a reasonable diff
if this situation occurs instead of crashing.

(cherry picked from commit 6338875b727ef947d0ccd608c5624b6ec5d5ba36)