Reject logins for non-interactive when envvar is unset
This change stops attempting to pop a web browser in non-interactive
sessions. Instead, the PULUMI_ACCESS_TOKEN environment variable must
be set. Otherwise, any attempt to use the CLI will yield

    $ pulumi --non-interactive preview
    error: PULUMI_ACCESS_TOKEN must be set for login during non-interactive CLI sessions

This is the behavior we want for Docker-based invocations of the CLI,
and so is part of pulumi/pulumi#1991.